pyproject.toml
[project]
name = "delivery-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Roman Skwortsoff",email = "roman.skwortsoff@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "ruff (>=0.12.8,<0.13.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pymongo (>=4.14.0,<5.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "pytest-dependency (>=0.6.0,<0.7.0)",
    "asgiref (>=3.9.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
